+++ src/components/board/PostList.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { getChannelPost } from "../../utils/api/getChannelPost";
 import { useLocation } from "react-router";
-import ImageCard from "../ImageCard";
+import ImageCard from "../common/ImageCard";
 
 // 무한스크롤에서 몇개씩 보여줄지 선택
 const limit = 12;
